BD-135: Database to Spark node Task-Url: https://knime-com.atlassian.net/browse/BD-135 BD-136: Spark to Database Task-Url: https://knime-com.atlassian.net/browse/BD-136 BD-163: Spark HDFS file reader/writer nodes Task-Url: https://knime-com.atlassian.net/browse/BD-163
Driver upload option disable by default and error handling for missing data source driver improved. We shouldn't upload new jars to the Spark job server which might conflict with existing installations to avoid version problems. If the driver is not available the node will complain with a descriptive error message that tells the user what to do.