BD-409: New Spark jobs plugin for Spark 1.6 on CDH 5.9 - 5.11 (adapted for releases/1.7)
The jobs only specifies model helpers and PMML port factory. JobRunFactories
and other extensions are provided by the existing Spark 1.6 jobs plugin.

It was necessary to refactor DefaultJobRunFactory, its variants as well as
all concrete JobRunFactories of existing jobs for this (resulting in removal
of a lot of boilerplate code)

BD-409: (Support Spark 1.6 on CDH 5.9 and above)