BD-807: disable transactions in Spark to Hive 2.3
Spark does not support tables created with transactions in Hive. Hive 3 enables transactions by default and HDP 3.0 too. This option was added in Hive 0.14.0 and should be compatible with HDP and CDH versions that support Spark 2.3.

BD-807 (Find out howto use shared warehouse dir for Spark and Hive on HDP 3.0.1)