BD-893: Set a global login user with simple auth in UGI
A global loginUser needs to be set in UGI, because otherwise code paths that
require Hadoop but not Kerberos (and hence don't use doAs()), will call
UGI.getLoginUser(), which may  attempt to do a Kerberos login itself and
spawn a ticket renewal thread.

When Kerberos is required, we need to use UGI.doAs() with a Kerberos
authenticated subject. If a library invokes UGI.getCurrentUser() inside
doAs(), this will use the subject from the AccessControlContext, not the
global loginUser.

BD-893: Hive Connector fails when executed on KNIME Server with embedded Hive JDBC driver