BD-942: Saves FlowVariables in node settings and applies their values during execution
For nodes in old workflows we extract the escaped name from the document and try to map them back to proper flow variable names. This can be ambiguous, in which case there will be warnings upon opening the dialog and execution.

BD-942 (PySpark nodes only update FlowVariables if Dialog is opened)

Approved-by: Björn Lohrmann <bjoern.lohrmann@knime.com>