BD-978: add username and password support to Livy
Add authentication options to use credentials/user/password to create a
Spark context on Livy. The user informations are added to the URL,
passed to Livy client and there extracted and used for HTTP basic
authentication. This is useful for KNOX or Azure HDInsight.

BD-978 (Support user name and password authentication in Create Spark
Context (Livy) node to connect to Azure HDInsight)