BD-817: use local hive scratch dir wrapper on windows
Add a wrapper around the local Hadoop file system on Windows and always
return the permission 0733 on the Hive scratch directory to avoid
confusing error messages if winutils.exe is unable to get the permission
from the real file system.

The file system wrapper reads the current scratch directory from the
configuration. It's important to close the file system together with the
Spark context, because the name of the scratch directory depends on the
node temp dir and changes on each execution.

BD-817 (Create Local Big Data Environment node fails to create root
    scratch directory)