Generate build.properties of spark local based on current included libs
On new files in the libs directory of Spark local, the plugin version gets only
bumped if the builds.properties changes. Add some task to generate the
bin.includes list, based on the jar and zip files, instead of an "libs/" entry,
to ensure the builds.properties changes.

See comments in AP-19650