BD-1268: add Guava to Local Spark
Hive requires Guava, but defines the dependency as provided. The Hive-JDBC driver bundled via bigdata-externals has an
explicit require-bundle, but the Hive version included in the libs directory of local Spark expects Guava to be on the
classpath. The old Hadoop 2.x client requires Guava too and implicit provided it, but the new Hadoop 3.x does not have
the dependency anymore. We now have to provide Guave manual in Local Spark.

BD-1268 (Update hadoop-client from 2.7.6 to 3.x)